requirements.txt
playwright>=1.49.0
gspread>=6.0.0
oauth2client>=4.1.3
matplotlib>=3.9.0
pandas>=2.2.0
requests>=2.32.0
feedparser>=6.0.10
yfinance>=0.2.36

python-dotenv>=1.0.0
